Changed Tooltip Styles and Improvements
Instead of default yellow, the tooltip colours are now tied to the hover colors and white. The FPS frame displays the tooltip again, and clicking on both the FPS and ping frame triggers the garbage collector. When under the specified minimum durability in the config, the armor icon and durability text now turns red. In the tooltip, any item under the minimum durability setting also turn red. Social icon's total online friends number is now closer and more centered to the icon.