Bugfixes, cleanup and documentation
Fixed a bug that caused the online friends number of the social and guild icon to display wrong when the XIV bar is set to be at the top instead of the bottom. Additionally fixed a bug that didn't properly hide friends that are not playing any game when that options is selected. Also cleaned up a bunch of code and added comments in SocialHover(), UpdateFriendText() and UpdateGuildText(). Added the profile option to offset the social/guild text along the y-axis.