Bugfix for SocialTooltip, Updates to GuildTooltip
Had a bug with realmName being nil when hovering over the social icon for the first time after logging in / reloading UI. This seems random and weird as fuck and I'm 90% sure this is some buggy fucky interaction on blizzard's end. Integrated an if-check to circumvent the error message. Nothing about this makes any sense.
Guild tooltip now shares its colouring with the rest of the addon's tooltips. Updated a deprecated API call to invite people to your party/raid. Members in the guild tooltip now no longer have their realm name attached to their name because that's fucking clutter.