Bugfix for new characters in 9.0.1 (no talents)
API call GetSpecialization() now returns 5 for newly created characters in shadowlands which causes an error because the specialization ID 5 does not exist for any spec. Not sure why they're not just returning nil for characters that don't yet have talents unlocked but it is how it is. This fix basically just disables the talent module if your low level char doesn't have talents unlocked yet.